pyproject.toml
[project]
name = "escrow-dex"
version = "0.1.0"
description = "Production-oriented FastAPI backend scaffold for an escrow freelance platform."
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "sqlalchemy>=2.0.32",
  "alembic>=1.13.2",
  "psycopg[binary]>=3.2.1",
  "pydantic>=2.8.2",
  "pydantic-settings>=2.4.0",
  "python-jose[cryptography]>=3.3.0",
  "passlib[argon2]>=1.7.4",
  "email-validator>=2.2.0",
  "celery[redis]>=5.4.0",
  "redis>=5.0.8",
  "structlog>=24.4.0",
  "httpx>=0.27.0"
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
include = ["app*"]
